+++ src/mlipaudit/benchmarks_cli.py
@@ -20,7 +20,7 @@ from dataclasses import replace
 from datetime import datetime
 from pathlib import Path
 
-from ase.calculators.calculator import Calculator as ASECalculator
+from ase.calculators.calculator import BaseCalculator as ASECalculator
 from mlip.models import Esen, ForceField, Mace, Nequip, Visnet
 from mlip.models.mlip_network import MLIPNetwork
 from mlip.models.model_io import load_model_from_zip